Add crypto::common::PublicKey. Move RrsigExt to dnssec::validator::base
