+++ src/crypto/common.rs
@@ -5,7 +5,9 @@
 //! documentation for each backend for more information.
 
 use core::fmt;
+use std::error;
 use std::sync::Arc;
+use std::vec::Vec;
 
 use ::ring::rand::SystemRandom;
 
@@ -13,6 +15,7 @@ use super::misc::{PublicKeyBytes, SignRaw, Signature};
 use crate::base::iana::SecAlg;
 use crate::dnssec::sign::error::{FromBytesError, GenerateError, SignError};
 use crate::dnssec::sign::SecretKeyBytes;
+use crate::rdata::Dnskey;
 
 #[cfg(feature = "openssl")]
 use super::openssl;
@@ -290,3 +293,107 @@ impl AsRef<[u8]> for Digest {
         }
     }
 }
+
+//----------- PublicKey ------------------------------------------------------
+
+pub enum PublicKey {
+    #[cfg(feature = "ring")]
+    Ring(ring::PublicKey),
+
+    #[cfg(feature = "openssl")]
+    Openssl(openssl::PublicKey),
+}
+
+impl PublicKey {
+    #[allow(unreachable_code)]
+    pub fn from_dnskey(
+        dnskey: &Dnskey<impl AsRef<[u8]>>,
+    ) -> Result<Self, AlgorithmError> {
+        #[cfg(feature = "ring")]
+        return Ok(Self::Ring(ring::PublicKey::from_dnskey(dnskey)?));
+
+        #[cfg(feature = "openssl")]
+        return Ok(Self::Openssl(openssl::PublicKey::from_dnskey(dnskey)?));
+
+        #[cfg(not(any(feature = "ring", feature = "openssl")))]
+        compile_error!("Either feature \"ring\" or \"openssl\" must be enabled for this crate.");
+    }
+
+    pub fn verify(
+        &self,
+        signed_data: &[u8],
+        signature: &[u8],
+    ) -> Result<(), AlgorithmError> {
+        match self {
+            #[cfg(feature = "ring")]
+            PublicKey::Ring(public_key) => {
+                public_key.verify(signed_data, signature)
+            }
+            #[cfg(feature = "openssl")]
+            PublicKey::Openssl(public_key) => {
+                public_key.verify(signed_data, signature)
+            }
+        }
+    }
+}
+
+/// Return the RSA exponent and modulus components from DNSKEY record data.
+pub fn rsa_exponent_modulus(
+    dnskey: &Dnskey<impl AsRef<[u8]>>,
+    min_len: usize,
+) -> Result<(Vec<u8>, Vec<u8>), AlgorithmError> {
+    let public_key = dnskey.public_key().as_ref();
+    if public_key.len() <= 3 {
+        return Err(AlgorithmError::InvalidData);
+    }
+
+    let (pos, exp_len) = match public_key[0] {
+        0 => (
+            3,
+            (usize::from(public_key[1]) << 8) | usize::from(public_key[2]),
+        ),
+        len => (1, usize::from(len)),
+    };
+
+    // Check if there's enough space for exponent and modulus.
+    if public_key[pos..].len() < pos + exp_len {
+        return Err(AlgorithmError::InvalidData);
+    };
+
+    // Check for minimum supported key size
+    if public_key[pos..].len() < min_len {
+        return Err(AlgorithmError::Unsupported);
+    }
+
+    let (e, n) = public_key[pos..].split_at(exp_len);
+    Ok((e.to_vec(), n.to_vec()))
+}
+
+//------------ AlgorithmError ------------------------------------------------
+
+/// An algorithm error during verification.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum AlgorithmError {
+    /// Unsupported algorithm.
+    Unsupported,
+
+    /// Bad signature.
+    BadSig,
+
+    /// Invalid data.
+    InvalidData,
+}
+
+//--- Display, Error
+
+impl fmt::Display for AlgorithmError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.write_str(match self {
+            AlgorithmError::Unsupported => "unsupported algorithm",
+            AlgorithmError::BadSig => "bad signature",
+            AlgorithmError::InvalidData => "invalid data",
+        })
+    }
+}
+
+impl error::Error for AlgorithmError {}
