+++ src/dnssec/validator/group.rs
@@ -18,9 +18,9 @@ use crate::base::opt::exterr::ExtendedError;
 use crate::base::rdata::ComposeRecordData;
 use crate::base::{Name, ParsedName, ParsedRecord, Record, Rtype, Ttl};
 use crate::crypto::common::{DigestContext, DigestType};
-use crate::crypto::validate::RrsigExt;
 use crate::dep::octseq::builder::with_infallible;
 use crate::dep::octseq::{Octets, OctetsFrom};
+use crate::dnssec::validator::base::RrsigExt;
 use crate::net::client::request::{RequestMessage, SendRequest};
 use crate::rdata::dnssec::Timestamp;
 use crate::rdata::{AllRecordData, Dnskey, Rrsig};
